Remove the ADIMemoryInterface and Memory structs and replace them with bare function calls.
Previously the ADIMemoryInterface and Memory structs added two additional layers to the probe-rs interface
this cluttered the code and made it less readable in general.
This commit effectively removes the two layer and should keep the functionality the same.

This is a first step towards a more facade like interface for the core which should make
future extensions as well as as the new custom debug sequences possible.
