+++ cli/src/info.rs
@@ -4,10 +4,10 @@ use probe_rs::{
     architecture::arm::{
         ap::{valid_access_ports, APClass, BaseaddrFormat, MemoryAP, BASE, BASE2, IDR},
         m0::Demcr,
-        memory::{ADIMemoryInterface, Component},
+        memory::{adi, Component},
         ArmCommunicationInterface, ArmCommunicationInterfaceState,
     },
-    CoreRegister, Memory,
+    CoreRegister,
 };
 
 use anyhow::Result;
@@ -67,23 +67,16 @@ pub(crate) fn show_info_of_device(shared_options: &SharedOptions) -> Result<()> 
                 };
                 baseaddr |= u64::from(base_register.BASEADDR << 12);
 
-                // We assume that only 32bit accesses work here, speed is not really an issue
-                let mut memory = Memory::new(ADIMemoryInterface::<ArmCommunicationInterface>::new(
-                    interface.reborrow(),
-                    access_port,
-                    true,
-                )?);
-
                 // Enable
                 // - Data Watchpoint and Trace (DWT)
                 // - Instrumentation Trace Macrocell (ITM)
                 // - Embedded Trace Macrocell (ETM)
                 // - Trace Port Interface Unit (TPIU).
-                let mut demcr = Demcr(memory.read_word_32(Demcr::ADDRESS)?);
+                let mut demcr = Demcr(adi::read_word_32(interface, access_port, Demcr::ADDRESS)?);
                 demcr.set_dwtena(true);
-                memory.write_word_32(Demcr::ADDRESS, demcr.into())?;
+                adi::write_word_32(interface, access_port, Demcr::ADDRESS, demcr.into())?;
 
-                let component_table = Component::try_parse(&mut memory, baseaddr as u64);
+                let component_table = Component::try_parse(interface.reborrow(), baseaddr as u64);
 
                 component_table
                     .iter()
