+++ probe-rs/src/architecture/arm/core/m4.rs
@@ -2,10 +2,10 @@ use crate::core::{
     CoreInformation, CoreInterface, CoreRegister, CoreRegisterAddress, RegisterFile,
 };
 use crate::error::Error;
-use crate::memory::Memory;
 use crate::DebugProbeError;
 
 use super::{register, reset_catch_clear, reset_catch_set, CortexState, Dfsr, ARM_REGISTER_FILE};
+use crate::architecture::arm::{memory::adi, ArmCommunicationInterface};
 use crate::{
     core::{Architecture, CoreStatus, HaltReason},
     MemoryInterface,
@@ -330,26 +330,34 @@ pub const MSP: CoreRegisterAddress = CoreRegisterAddress(0b000_1001);
 pub const PSP: CoreRegisterAddress = CoreRegisterAddress(0b000_1010);
 
 pub struct M4<'probe> {
-    memory: Memory<'probe>,
+    interface: ArmCommunicationInterface<'probe>,
 
     state: &'probe mut CortexState,
 }
 
 impl<'probe> M4<'probe> {
     pub(crate) fn new(
-        mut memory: Memory<'probe>,
+        mut interface: ArmCommunicationInterface<'probe>,
         state: &'probe mut CortexState,
     ) -> Result<M4<'probe>, Error> {
         if !state.initialized() {
             // determine current state
-            let dhcsr = Dhcsr(memory.read_word_32(Dhcsr::ADDRESS)?);
+            let dhcsr = Dhcsr(adi::read_word_32(
+                &mut interface,
+                state.memory_ap,
+                Dhcsr::ADDRESS,
+            )?);
 
             let core_state = if dhcsr.s_sleep() {
                 CoreStatus::Sleeping
             } else if dhcsr.s_halt() {
                 log::debug!("Core was halted when connecting");
 
-                let dfsr = Dfsr(memory.read_word_32(Dfsr::ADDRESS)?);
+                let dfsr = Dfsr(adi::read_word_32(
+                    &mut interface,
+                    state.memory_ap,
+                    Dfsr::ADDRESS,
+                )?);
 
                 let reason = dfsr.halt_reason();
 
@@ -362,13 +370,18 @@ impl<'probe> M4<'probe> {
             // so we clear them here to ensure that that none are set.
             let dfsr_clear = Dfsr::clear_all();
 
-            memory.write_word_32(Dfsr::ADDRESS, dfsr_clear.into())?;
+            adi::write_word_32(
+                &mut interface,
+                state.memory_ap,
+                Dfsr::ADDRESS,
+                dfsr_clear.into(),
+            )?;
 
             state.current_state = core_state;
-            state.initialize();
+            state.initialize(&mut interface)?;
         }
 
-        Ok(Self { memory, state })
+        Ok(Self { interface, state })
     }
 
     fn wait_for_core_register_transfer(&mut self, timeout: Duration) -> Result<(), Error> {
@@ -377,7 +390,7 @@ impl<'probe> M4<'probe> {
         let start = Instant::now();
 
         while start.elapsed() < timeout {
-            let dhcsr_val = Dhcsr(self.memory.read_word_32(Dhcsr::ADDRESS)?);
+            let dhcsr_val = Dhcsr(self.read_word_32(Dhcsr::ADDRESS)?);
 
             if dhcsr_val.s_regrdy() {
                 return Ok(());
@@ -393,7 +406,7 @@ impl<'probe> CoreInterface for M4<'probe> {
         let start = Instant::now();
 
         while start.elapsed() < timeout {
-            let dhcsr_val = Dhcsr(self.memory.read_word_32(Dhcsr::ADDRESS)?);
+            let dhcsr_val = Dhcsr(self.read_word_32(Dhcsr::ADDRESS)?);
             if dhcsr_val.s_halt() {
                 // update halted state
                 self.status()?;
@@ -406,7 +419,7 @@ impl<'probe> CoreInterface for M4<'probe> {
 
     fn core_halted(&mut self) -> Result<bool, Error> {
         // Wait until halted state is active again.
-        let dhcsr_val = Dhcsr(self.memory.read_word_32(Dhcsr::ADDRESS)?);
+        let dhcsr_val = Dhcsr(self.read_word_32(Dhcsr::ADDRESS)?);
 
         if dhcsr_val.s_halt() {
             Ok(true)
@@ -416,7 +429,7 @@ impl<'probe> CoreInterface for M4<'probe> {
     }
 
     fn status(&mut self) -> Result<CoreStatus, Error> {
-        let dhcsr = Dhcsr(self.memory.read_word_32(Dhcsr::ADDRESS)?);
+        let dhcsr = Dhcsr(self.read_word_32(Dhcsr::ADDRESS)?);
 
         if dhcsr.s_sleep() {
             // Check if we assumed the core to be halted
@@ -432,13 +445,12 @@ impl<'probe> CoreInterface for M4<'probe> {
         // TODO: Handle lockup
 
         if dhcsr.s_halt() {
-            let dfsr = Dfsr(self.memory.read_word_32(Dfsr::ADDRESS)?);
+            let dfsr = Dfsr(self.read_word_32(Dfsr::ADDRESS)?);
 
             let reason = dfsr.halt_reason();
 
             // Clear bits from Dfsr register
-            self.memory
-                .write_word_32(Dfsr::ADDRESS, Dfsr::clear_all().into())?;
+            self.write_word_32(Dfsr::ADDRESS, Dfsr::clear_all().into())?;
 
             // If the core was halted before, we cannot read the halt reason from the chip,
             // because we clear it directly after reading.
@@ -478,17 +490,15 @@ impl<'probe> CoreInterface for M4<'probe> {
         dcrsr_val.set_regwnr(false); // Perform a read.
         dcrsr_val.set_regsel(addr.into()); // The address of the register to read.
 
-        self.memory
-            .write_word_32(Dcrsr::ADDRESS, dcrsr_val.into())?;
+        self.write_word_32(Dcrsr::ADDRESS, dcrsr_val.into())?;
 
         self.wait_for_core_register_transfer(Duration::from_millis(100))?;
 
-        self.memory.read_word_32(Dcrdr::ADDRESS).map_err(From::from)
+        self.read_word_32(Dcrdr::ADDRESS).map_err(From::from)
     }
 
     fn write_core_reg(&mut self, addr: CoreRegisterAddress, value: u32) -> Result<()> {
         let result: Result<(), Error> = self
-            .memory
             .write_word_32(Dcrdr::ADDRESS, value)
             .map_err(From::from);
         result?;
@@ -498,8 +508,7 @@ impl<'probe> CoreInterface for M4<'probe> {
         dcrsr_val.set_regwnr(true); // Perform a write.
         dcrsr_val.set_regsel(addr.into()); // The address of the register to write.
 
-        self.memory
-            .write_word_32(Dcrsr::ADDRESS, dcrsr_val.into())?;
+        self.write_word_32(Dcrsr::ADDRESS, dcrsr_val.into())?;
 
         self.wait_for_core_register_transfer(Duration::from_millis(100))
             .context("Waiting for core register transfer")
@@ -513,7 +522,7 @@ impl<'probe> CoreInterface for M4<'probe> {
         value.set_c_debugen(true);
         value.enable_write();
 
-        self.memory.write_word_32(Dhcsr::ADDRESS, value.into())?;
+        self.write_word_32(Dhcsr::ADDRESS, value.into())?;
 
         self.wait_for_core_halted(timeout)?;
 
@@ -530,8 +539,8 @@ impl<'probe> CoreInterface for M4<'probe> {
         value.set_c_debugen(true);
         value.enable_write();
 
-        self.memory.write_word_32(Dhcsr::ADDRESS, value.into())?;
-        self.memory.flush()?;
+        self.write_word_32(Dhcsr::ADDRESS, value.into())?;
+        self.flush()?;
 
         // We assume that the core is running now
         self.state.current_state = CoreStatus::Running;
@@ -549,7 +558,7 @@ impl<'probe> CoreInterface for M4<'probe> {
         value.set_c_maskints(true);
         value.enable_write();
 
-        self.memory.write_word_32(Dhcsr::ADDRESS, value.into())?;
+        self.write_word_32(Dhcsr::ADDRESS, value.into())?;
 
         self.wait_for_core_halted(Duration::from_millis(100))?;
 
@@ -566,7 +575,7 @@ impl<'probe> CoreInterface for M4<'probe> {
         value.vectkey();
         value.set_sysresetreq(true);
 
-        self.memory.write_word_32(Aircr::ADDRESS, value.into())?;
+        self.write_word_32(Aircr::ADDRESS, value.into())?;
 
         Ok(())
     }
@@ -596,7 +605,7 @@ impl<'probe> CoreInterface for M4<'probe> {
     }
 
     fn get_available_breakpoint_units(&mut self) -> Result<u32, Error> {
-        let raw_val = self.memory.read_word_32(FpCtrl::ADDRESS)?;
+        let raw_val = self.read_word_32(FpCtrl::ADDRESS)?;
 
         let reg = FpCtrl::from(raw_val);
 
@@ -613,7 +622,7 @@ impl<'probe> CoreInterface for M4<'probe> {
         val.set_key(true);
         val.set_enable(state);
 
-        self.memory.write_word_32(FpCtrl::ADDRESS, val.into())?;
+        self.write_word_32(FpCtrl::ADDRESS, val.into())?;
 
         self.state.hw_breakpoints_enabled = true;
 
@@ -621,7 +630,7 @@ impl<'probe> CoreInterface for M4<'probe> {
     }
 
     fn set_breakpoint(&mut self, bp_unit_index: usize, addr: u32) -> Result<(), Error> {
-        let raw_val = self.memory.read_word_32(FpCtrl::ADDRESS)?;
+        let raw_val = self.read_word_32(FpCtrl::ADDRESS)?;
         let ctrl_reg = FpCtrl::from(raw_val);
 
         let val: u32;
@@ -639,7 +648,7 @@ impl<'probe> CoreInterface for M4<'probe> {
         // address spaces than Rev1.
         let reg_addr = FpRev1CompX::ADDRESS + (bp_unit_index * size_of::<u32>()) as u32;
 
-        self.memory.write_word_32(reg_addr, val)?;
+        self.write_word_32(reg_addr, val)?;
 
         Ok(())
     }
@@ -654,7 +663,7 @@ impl<'probe> CoreInterface for M4<'probe> {
 
         let reg_addr = FpRev1CompX::ADDRESS + (bp_unit_index * size_of::<u32>()) as u32;
 
-        self.memory.write_word_32(reg_addr, val.into())?;
+        self.write_word_32(reg_addr, val.into())?;
 
         Ok(())
     }
@@ -670,31 +679,51 @@ impl<'probe> CoreInterface for M4<'probe> {
 
 impl<'probe> MemoryInterface for M4<'probe> {
     fn read_word_32(&mut self, address: u32) -> Result<u32, Error> {
-        self.memory.read_word_32(address)
+        adi::read_word_32(&mut self.interface, self.state.memory_ap, address).map_err(From::from)
     }
+
     fn read_word_8(&mut self, address: u32) -> Result<u8, Error> {
-        self.memory.read_word_8(address)
+        if self.state.supports_only_32bit_access {
+            adi::read_word_8(&mut self.interface, self.state.memory_ap, address).map_err(From::from)
+        } else {
+            adi::read_word_8_true(&mut self.interface, self.state.memory_ap, address)
+                .map_err(From::from)
+        }
     }
+
     fn read_32(&mut self, address: u32, data: &mut [u32]) -> Result<(), Error> {
-        self.memory.read_32(address, data)
+        adi::read_32(&mut self.interface, self.state.memory_ap, address, data).map_err(From::from)
     }
+
     fn read_8(&mut self, address: u32, data: &mut [u8]) -> Result<(), Error> {
-        self.memory.read_8(address, data)
+        adi::read_8(&mut self.interface, self.state.memory_ap, address, data).map_err(From::from)
     }
+
     fn write_word_32(&mut self, address: u32, data: u32) -> Result<(), Error> {
-        self.memory.write_word_32(address, data)
+        adi::write_word_32(&mut self.interface, self.state.memory_ap, address, data)
+            .map_err(From::from)
     }
+
     fn write_word_8(&mut self, address: u32, data: u8) -> Result<(), Error> {
-        self.memory.write_word_8(address, data)
+        if self.state.supports_only_32bit_access {
+            adi::write_word_8(&mut self.interface, self.state.memory_ap, address, data)
+                .map_err(From::from)
+        } else {
+            adi::write_word_8_true(&mut self.interface, self.state.memory_ap, address, data)
+                .map_err(From::from)
+        }
     }
+
     fn write_32(&mut self, address: u32, data: &[u32]) -> Result<(), Error> {
-        self.memory.write_32(address, data)
+        adi::write_32(&mut self.interface, self.state.memory_ap, address, data).map_err(From::from)
     }
+
     fn write_8(&mut self, address: u32, data: &[u8]) -> Result<(), Error> {
-        self.memory.write_8(address, data)
+        adi::write_8(&mut self.interface, self.state.memory_ap, address, data).map_err(From::from)
     }
+
     fn flush(&mut self) -> Result<(), Error> {
-        self.memory.flush()
+        adi::flush(&mut self.interface).map_err(From::from)
     }
 }
