+++ probe-rs/src/architecture/arm/memory/mod.rs
@@ -1,8 +1,7 @@
-pub(crate) mod adi_v5_memory_interface;
+pub mod adi;
 pub(crate) mod romtable;
 
 use super::ap::AccessPortError;
-pub use adi_v5_memory_interface::ADIMemoryInterface;
 pub use romtable::Component;
 
 pub trait ToMemoryReadSize: Into<u32> + Copy {
